+++ packages/insomnia/src/runtimes/crypto/crypt-adapter.renderer.ts
@@ -1,4 +1,4 @@
-import type { AESMessage } from '../account/crypt';
+import type { AESMessage } from '../../account/crypt';
 
 export const decryptAES = (symmetricKey: string | JsonWebKey, encryptedResult: AESMessage): Promise<string> =>
   window.main.crypt.decryptAES(symmetricKey, encryptedResult);
