+++ packages/insomnia/src/runtimes/runtime.node.ts
@@ -1,0 +1,12 @@
+import * as networkAdapter from './network/network-adapter.node';
+import * as renderAdapter from './templating/render-adapter.node';
+import * as cryptAdapter from './crypto/crypt-adapter.node';
+import * as secretStorageAdapter from './secret-storage/secret-storage-adapter.node';
+import type { RuntimeCapabilities } from './types';
+
+export const nodeRuntime = {
+  network: networkAdapter,
+  crypto: cryptAdapter,
+  templating: renderAdapter,
+  secretStorage: secretStorageAdapter,
+} satisfies RuntimeCapabilities;
