+++ packages/insomnia/src/runtimes/templating/render-adapter.node.ts
@@ -1,6 +1,6 @@
-import type { RenderInputType } from './types';
+import type { RenderInputType } from '../../templating/types';
 
 export async function renderTemplate({ input, context, path, ignoreUndefinedEnvVariable }: RenderInputType): Promise<string | null> {
-  const templating = await import('./index');
+  const templating = await import('../../templating/index');
   return templating.render(input, { context, path, ignoreUndefinedEnvVariable });
 }
