+++ packages/insomnia/src/runtimes/templating/render-adapter.renderer.ts
@@ -1,0 +1,6 @@
+import type { RenderInputType } from '../../templating/types';
+
+export async function renderTemplate(input: RenderInputType): Promise<string | null> {
+  const { renderInWorker } = await import('../../ui/worker/templating-handler');
+  return renderInWorker(input);
+}
