+++ packages/insomnia/src/runtimes/templating/render-adapter.ts
