+++ packages/insomnia/src/runtimes/types.ts
@@ -1,8 +1,8 @@
 import type { AESMessage, Cookie, RequestHeader } from 'insomnia-data';
 
-import type { RequestContext } from '../../../../insomnia-scripting-environment/src/objects';
-import type { CurlRequestOptions, CurlRequestOutput, ResponsePatch } from '../../main/network/libcurl-promise';
-import type { RenderedRequest, RenderInputType } from '../../templating/types';
+import type { RequestContext } from '../../../insomnia-scripting-environment/src/objects';
+import type { CurlRequestOptions, CurlRequestOutput, ResponsePatch } from '../main/network/libcurl-promise';
+import type { RenderedRequest, RenderInputType } from '../templating/types';
 
 interface CurlRequestErrorOutput {
   statusMessage: string;
