--- packages/insomnia/src/templating/render-adapter.renderer.ts
@@ -1,6 +1,0 @@
-import type { RenderInputType } from './types';
-
-export async function renderTemplate(input: RenderInputType): Promise<string | null> {
-  const { renderInWorker } = await import('../ui/worker/templating-handler');
-  return renderInWorker(input);
-}
