+++ packages/insomnia/src/ui/components/editors/environment-key-value-editor/key-value-editor.tsx
@@ -19,7 +19,7 @@ import { OneLineEditor } from '~/ui/components/.client/codemirror/one-line-edito
 import { checkNestedKeys, ensureKeyIsValid } from '~/utils/environment-utils';
 
 import { generateId } from '../../../../common/misc';
-import { decryptSecretValue, encryptSecretValue } from '../../../../utils/crypt-adapter';
+import { decryptSecretValue, encryptSecretValue } from '~/runtimes/crypto/crypt-adapter';
 import { base64decode } from '../../../../utils/vault';
 import { PromptButton } from '../../base/prompt-button';
 import { Icon } from '../../icon';
